[test] fix issue7305-nwfaketop-cookie: kill server on windows

+++ test/sanity/issue7305-nwfaketop-cookie/test.py
@@ -50,5 +50,9 @@ try:
     assert (svrlog == expected)
 
 finally:
-    server.terminate()
+    import platform
+    if platform.system() == 'Windows':
+        subprocess.call(['taskkill', '/F', '/T', '/PID', str(server.pid)])
+    else:
+        server.terminate()
     driver.quit()
